fix: fix macOS and FreeBSD CI build errors
timer_osx.cpp: uint64_t idle time divided by 1e6 must be explicitly
cast to uint32_t before returning; the conversion is safe since
idle-time in ms easily fits 32 bits.

cameradevice.cpp: toCharArray() used a signed 'int' cursor to index a
std::array, triggering -Werror,-Wsign-conversion on macOS clang.
Change the cursor to size_t with an explicit cast from numDigits().

freebsd.yml: pkg create -r rootdir prepends rootdir + prefix to each
PLIST entry, so rootdir must be '_pkg' (not '_pkg/usr/local'); passing
'_pkg/usr/local' caused a double usr/local in the lookup path.

+++ src/platform/timer_osx.cpp
@@ -54,7 +54,7 @@ uint32_t Platform::getIdleTime()
     CFNumberGetValue(static_cast<CFNumberRef>(property), kCFNumberSInt64Type, &idleTime_ns);
     CFRelease(property);
 
-    return idleTime_ns / 1000000;
+    return static_cast<uint32_t>(idleTime_ns / 1000000U);
 }
 #endif
 #endif
